+++ R/checkInputs.R
@@ -43,7 +43,7 @@
 
 ##################################################################################
 
-.checkObstacles = function(obstacles, obstacles_height_field, messages) {
+.checkObstacles = function(obstacles, obstacles_height_field) {
 
   # Check projected
   if(!is.na(proj4string(obstacles)) & !sp::is.projected(obstacles))
@@ -61,18 +61,58 @@
   if(any(obstacles@data[, obstacles_height_field] < 0))
     stop("'obstacles_height_field' cannot contain negative values")
 
-  # Print units assumption
-  if(messages) {
-    message(
-      paste0(
-        "Assuming ", obstacles_height_field, " given in projection units (",
-        gsub(" .*", "",
-             gsub(".*\\+units=", "", sp::proj4string(obstacles))
-        ),
-        ")"
-      )
-    )
-  }
+  # # Print units assumption
+  # if(messages) {
+  #   message(
+  #     paste0(
+  #       "Assuming ", obstacles_height_field, " given in projection units (",
+  #       gsub(" .*", "",
+  #            gsub(".*\\+units=", "", sp::proj4string(obstacles))
+  #       ),
+  #       ")"
+  #     )
+  #   )
+  # }
+}
+
+##################################################################################
+
+.checkSolarRad = function(solar_normal, solar_diffuse, solar_pos) {
+
+  if(!is.vector(solar_normal) | !is.numeric(solar_normal))
+    stop("'solar_normal' must be a numeric vector")
+
+  if(!is.vector(solar_diffuse) | !is.numeric(solar_diffuse))
+    stop("'solar_diffuse' must be a numeric vector")
+
+  if(!all.equal(length(solar_normal), length(solar_diffuse), nrow(solar_pos)))
+    stop("'solar_normal', 'solar_diffuse' and 'solar_pos' must have same length")
+
+}
+
+##################################################################################
+
+.checkGrid = function(grid) {
+
+  if(class(grid) != c("SpatialPointsDataFrame"))
+    stop("'grid' must be 'SpatialPointsDataFrame'")
+
+  if(dimensions(grid) != 3)
+    stop("'grid' must be three-dimensional")
+
+  if(!"type" %in% colnames(grid@data))
+    stop("'grid' must contain an attribute named 'type'")
+
+  if(!"facade_az" %in% colnames(grid@data))
+    stop("'grid' must contain an attribute named 'facade_az'")
+
+  # Azimuth in [0,360]
+  if(any(!is.na(grid$facade_az) & (grid$facade_az < 0 | grid$facade_az > 360)))
+    stop("'grid$facade_az' should be numeric in [0, 360]")
+
+  if(length(setdiff(grid$type, c("roof", "facade"))) > 0)
+    stop("All values of 'type' attribute must be 'roof' or 'facade'")
+
 }
 
 ##################################################################################